Store a summary of symbols by letter on each line
This allows us to skip a lot of symbols, thus 
making the iteration way faster. Moreover, we can
keep splicing into the SymbolStore fast because as
soon as lines get removed, the index is
automatically kept in-sync (e.g. because it's
stored per each line). On a file with ~ 26k lines,
autocompleting a word takes < 80ms.