Force redraw in order to avoid display bug
The highlighted background doesn't always go all the way to the end and it's driving me crazy. It's actually a bug with Chrome where it doesn't properly update. If in the dev tools you remove and add some properties again it goes back to normal.

This is a hacky fix but that works. Let me know if you have better ideas!

In https://github.com/facebook/nuclide/blob/master/pkg/nuclide-file-tree/components/FileTree.js#L54 type `this.fixn` and make sure the background in displayed correctly.

Before:
![](http://g.recordit.co/neDExfom91.gif)

After:
![](http://g.recordit.co/NQNGoY1cfH.gif)