Make specs asynchronous
The `waitForAutocomplete()` method used to wait an arbitrary amount of 
miliseconds, which caused build flakiness on Atom due to race 
conditions. In order to fix that, `waitForAutocomplete()` now is an 
async method that actually waits until the autocomplete modal is shown. 
In order to use it, we've needed to refactor all the tests that used 
this method to be async.

Co-Authored-By: Nathan Sobo <nathan@github.com>